clean-up and reorganization
