+++ test/aqua.jl
@@ -1,11 +1,7 @@
-module AquaVectorInterface
-
 using Test
 using VectorInterface
 using Aqua
 
 @testset "Aqua" begin
     Aqua.test_all(VectorInterface)
-end
-
 end
