+++ test/complicated.jl
@@ -1,13 +1,20 @@
-module Simple
 using VectorInterface
 using Test
 using TestExtras
 
+using VectorInterface: One
+
 deepcollect(x) = vcat(map(deepcollect, x)...)
 deepcollect(x::Number) = x
 
-x = randn(3, 3, 3)
-y = randn(3, 3, 3)
+x = (
+    NamedTuple{(:x, :y)}.(collect(zip(randn(2, 2), rand(2, 2)))),
+    (randn(), randn(3), randn(2, 2)'), randn(), (view(randn(4, 4), 1:2, [1, 3, 4]),),
+)
+y = (
+    NamedTuple{(:x, :y)}.(collect(zip(randn(2, 2), rand(2, 2)))),
+    (randn(), randn(3), randn(2, 2)'), randn(), (view(randn(4, 4), 1:2, [1, 3, 4]),),
+)
 
 @testset "scalartype" begin
     s = @constinferred scalartype(x)
@@ -22,34 +29,40 @@ end
     @test all(deepcollect(z) .=== zero(scalartype(x)))
     z1 = @constinferred zerovector!!(deepcopy(x))
     @test all(deepcollect(z1) .=== zero(scalartype(x)))
-    z2 = @constinferred zerovector!(deepcopy(x))
-    @test all(deepcollect(z2) .=== zero(scalartype(x)))
 
-    z3 = @constinferred zerovector(x, ComplexF64)
+    z2 = @constinferred zerovector(x, ComplexF64)
+    @test all(deepcollect(z2) .=== zero(ComplexF64))
+    z3 = @constinferred zerovector!!(deepcopy(x), ComplexF64)
     @test all(deepcollect(z3) .=== zero(ComplexF64))
-    z4 = @constinferred zerovector!!(deepcopy(x), ComplexF64)
-    @test all(deepcollect(z4) .=== zero(ComplexF64))
-    @test_throws MethodError zerovector!(deepcopy(x), ComplexF64)
+
+    xy = [deepcopy(x), deepcopy(y)]
+    z4 = @constinferred zerovector!(xy)
+    @test all(deepcollect(z4) .=== zero(scalartype(xy)))
+    @test_throws MethodError zerovector!(xy, ComplexF64)
 end
 
 @testset "scale" begin
     α = randn()
     z = @constinferred scale(x, α)
     @test all(deepcollect(z) .== α .* deepcollect(x))
-
+    z = @constinferred scale(x, One())
+    @test all(deepcollect(z) .== deepcollect(x))
     z2 = @constinferred scale!!(deepcopy(x), α)
     @test deepcollect(z2) ≈ (α .* deepcollect(x))
+    z2 = @constinferred scale!!(deepcopy(x), One())
+    @test all(deepcollect(z2) .== deepcollect(x))
     xcopy = deepcopy(x)
     z2 = @constinferred scale!!(deepcopy(y), xcopy, α)
     @test deepcollect(z2) ≈ (α .* deepcollect(x))
     @test all(deepcollect(xcopy) .== deepcollect(x))
 
-    z3 = @constinferred scale!(deepcopy(x), α)
-    @test deepcollect(z3) ≈ (α .* deepcollect(x))
-    xcopy = deepcopy(x)
-    z3 = @constinferred scale!(zerovector(x), xcopy, α)
-    @test deepcollect(z3) ≈ (α .* deepcollect(x))
-    @test all(deepcollect(xcopy) .== deepcollect(x))
+    xy = [deepcopy(x), deepcopy(y)]
+    z3 = @constinferred scale!(deepcopy(xy), α)
+    @test deepcollect(z3) ≈ (α .* deepcollect(xy))
+    xycopy = deepcopy(xy)
+    z3 = @constinferred scale!(zerovector(xy), xycopy, α)
+    @test deepcollect(z3) ≈ (α .* deepcollect(xy))
+    @test all(deepcollect(xycopy) .== deepcollect(xy))
 
     α = randn(ComplexF64)
     z4 = @constinferred scale(x, α)
@@ -58,30 +71,16 @@ end
     z5 = @constinferred scale!!(xcopy, α)
     @test deepcollect(z5) ≈ (α .* deepcollect(x))
     @test all(deepcollect(xcopy) .== deepcollect(x))
-    @test_throws InexactError scale!(xcopy, α)
 
     α = randn(ComplexF64)
     xcopy = deepcopy(x)
     z6 = @constinferred scale!!(zerovector(x), xcopy, α)
     @test deepcollect(z6) ≈ (α .* deepcollect(x))
     @test all(deepcollect(xcopy) .== deepcollect(x))
-    @test_throws InexactError scale!(zerovector(x), xcopy, α)
-
     xz = @constinferred zerovector(x, ComplexF64)
     z6 = @constinferred scale!!(xz, xcopy, α)
     @test deepcollect(z6) ≈ (α .* deepcollect(x))
     @test all(deepcollect(xcopy) .== deepcollect(x))
-
-    z7 = @constinferred scale!(xz, xcopy, α)
-    @test deepcollect(z7) ≈ (α .* deepcollect(x))
-    @test all(deepcollect(xcopy) .== deepcollect(x))
-
-    ycomplex = zerovector(y, ComplexF64)
-    α = randn(Float64)
-    xcopy = deepcopy(x)
-    z8 = @constinferred scale!!(ycomplex, xcopy, α)
-    @test z8 === ycomplex
-    @test all(deepcollect(z8) .== α .* deepcollect(xcopy))
 end
 
 @testset "add" begin
@@ -91,36 +90,50 @@ end
     z = @constinferred add(y, x, α)
     # for some reason, on some Julia versions on some platforms, but only in test mode
     # there is a small floating point discrepancy, which makes the following test fail:
-    # @test all(deepcollect(z) .== muladd.(deepcollect(x), α, deepcollect(y)))
+    # @test all(deepcollect(z) ≈ muladd.(deepcollect(x), α, deepcollect(y)))
     @test deepcollect(z) ≈ muladd.(deepcollect(x), α, deepcollect(y))
+
     z = @constinferred add(y, x, α, β)
     # for some reason, on some Julia versions on some platforms, but only in test mode
     # there is a small floating point discrepancy, which makes the following test fail:
     # @test all(deepcollect(z) .== muladd.(deepcollect(x), α, deepcollect(y) .* β))
     @test deepcollect(z) ≈ muladd.(deepcollect(x), α, deepcollect(y) .* β)
 
-    α, β = randn(2)
-    xcopy = deepcopy(x)
-    z2 = @constinferred add!!(deepcopy(y), xcopy)
-    @test deepcollect(z2) ≈ (deepcollect(x) .+ deepcollect(y))
-    @test all(deepcollect(xcopy) .== deepcollect(x))
-    z2 = @constinferred add!!(deepcopy(y), xcopy, α)
-    @test deepcollect(z2) ≈ (muladd.(deepcollect(x), α, deepcollect(y)))
-    @test all(deepcollect(xcopy) .== deepcollect(x))
-    z2 = @constinferred add!!(deepcopy(y), xcopy, α, β)
-    @test deepcollect(z2) ≈ (muladd.(deepcollect(x), α, deepcollect(y) .* β))
-    @test all(deepcollect(xcopy) .== deepcollect(x))
+    z = @constinferred add(y, x, One(), β)
+    @test deepcollect(z) ≈ muladd.(deepcollect(y), β, deepcollect(x))
+    z = @constinferred add(y, x, α, One())
+    @test deepcollect(z) ≈ muladd.(deepcollect(x), α, deepcollect(y))
+    z = @constinferred add(y, x, One(), One())
+    @test all(deepcollect(z) .== deepcollect(x) .+ deepcollect(y))
+    z = @constinferred add(y, x, One())
+    @test all(deepcollect(z) .== deepcollect(x) .+ deepcollect(y))
 
     α, β = randn(2)
-    z3 = @constinferred add!(deepcopy(y), xcopy)
-    @test deepcollect(z3) ≈ (deepcollect(y) .+ deepcollect(x))
-    @test all(deepcollect(xcopy) .== deepcollect(x))
-    z3 = @constinferred add!(deepcopy(y), xcopy, α)
-    @test all(deepcollect(xcopy) .== deepcollect(x))
-    @test deepcollect(z3) ≈ (muladd.(deepcollect(x), α, deepcollect(y)))
-    z3 = @constinferred add!(deepcopy(y), xcopy, α, β)
-    @test deepcollect(z3) ≈ (muladd.(deepcollect(x), α, deepcollect(y) .* β))
-    @test all(deepcollect(xcopy) .== deepcollect(x))
+    z2 = @constinferred add!!(deepcopy(y), deepcopy(x))
+    @test deepcollect(z2) ≈ (deepcollect(x) .+ deepcollect(y))
+    z2 = @constinferred add!!(deepcopy(y), deepcopy(x), α)
+    @test deepcollect(z2) ≈ (muladd.(deepcollect(x), α, deepcollect(y)))
+    z2 = @constinferred add!!(deepcopy(y), deepcopy(x), α, β)
+    @test deepcollect(z2) ≈ (muladd.(deepcollect(x), α, deepcollect(y) .* β))
+
+    z2 = @constinferred add!!(deepcopy(y), deepcopy(x), One(), β)
+    @test deepcollect(z2) ≈ muladd.(deepcollect(y), β, deepcollect(x))
+    z2 = @constinferred add!!(deepcopy(y), deepcopy(x), α, One())
+    @test deepcollect(z2) ≈ muladd.(deepcollect(x), α, deepcollect(y))
+    z2 = @constinferred add!!(deepcopy(y), deepcopy(x), One(), One())
+    @test all(deepcollect(z2) .== deepcollect(x) .+ deepcollect(y))
+    z2 = @constinferred add!!(deepcopy(y), deepcopy(x), One())
+    @test all(deepcollect(z2) .== deepcollect(x) .+ deepcollect(y))
+
+    α, β = randn(2)
+    xy = [deepcopy(x), deepcopy(y)]
+    yx = [deepcopy(y), deepcopy(x)]
+    z3 = @constinferred add!(deepcopy(xy), deepcopy(yx))
+    @test deepcollect(z3) ≈ (deepcollect(xy) .+ deepcollect(yx))
+    z3 = @constinferred add!(deepcopy(xy), deepcopy(yx), α)
+    @test deepcollect(z3) ≈ (muladd.(deepcollect(yx), α, deepcollect(xy)))
+    z3 = @constinferred add!(deepcopy(xy), deepcopy(yx), α, β)
+    @test deepcollect(z3) ≈ (muladd.(deepcollect(yx), α, deepcollect(xy) .* β))
 
     α, β = randn(ComplexF64, 2)
     z4 = @constinferred add(y, x, α)
@@ -129,24 +142,10 @@ end
     @test deepcollect(z4) ≈ (muladd.(deepcollect(x), α, deepcollect(y) .* β))
 
     α, β = randn(ComplexF64, 2)
-    z5 = @constinferred add!!(deepcopy(y), xcopy, α)
+    z5 = @constinferred add!!(deepcopy(y), deepcopy(x), α)
     @test deepcollect(z5) ≈ (muladd.(deepcollect(x), α, deepcollect(y)))
-    @test all(deepcollect(xcopy) .== deepcollect(x))
-    z5 = @constinferred add!!(deepcopy(y), xcopy, α, β)
+    z5 = @constinferred add!!(deepcopy(y), deepcopy(x), α, β)
     @test deepcollect(z5) ≈ (muladd.(deepcollect(x), α, deepcollect(y) .* β))
-    @test all(deepcollect(xcopy) .== deepcollect(x))
-
-    α, β = randn(ComplexF64, 2)
-    z5 = @constinferred add!!(deepcopy(y), xcopy, α)
-    @test deepcollect(z5) ≈ (muladd.(deepcollect(x), α, deepcollect(y)))
-    @test all(deepcollect(xcopy) .== deepcollect(x))
-    z5 = @constinferred add!!(deepcopy(y), xcopy, α, β)
-    @test deepcollect(z5) ≈ (muladd.(deepcollect(x), α, deepcollect(y) .* β))
-    @test all(deepcollect(xcopy) .== deepcollect(x))
-
-    α, β = randn(ComplexF64, 2)
-    @test_throws InexactError add!(deepcopy(y), xcopy, α)
-    @test_throws InexactError add!(deepcopy(y), xcopy, α, β)
 end
 
 @testset "inner" begin
@@ -156,6 +155,4 @@ end
     α, β = randn(ComplexF64, 2)
     s2 = @constinferred inner(scale(x, α), scale(y, β))
     @test s2 ≈ inner(α * deepcollect(x), β * deepcollect(y))
-end
-
 end
